Update citeste_bonul.py
The program now stores the product name and its price in a tuple, and then prints it.

To test this, I used a receipt from Lidl. After cropping it to remove any personal or irrelevant information, I printed the text. Here's an example of how a product appears:

2.000
BUC
X 3 , 99
Sardine
in
sos
tomat
7, 98

In this example, there are 2 pieces of sardines, each costing 3.99 (2.000 BUC x 3 , 99), which is information I don’t need.

What I do need is the product name "Sardine in sos tomat" which is spread across several lines, followed by the total price, "7, 98". This is the data I want to extract and store.

I've also decided to keep the total price listed at the end of the receipt.

There are still some exceptions I need to account for, such as product names that may contain numbers or special characters.